Move js-tequilapi production sources to src
