+++ src/connection/connect-count-dto.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2017 The "mysteriumnetwork/js-tequilapi" Authors.
+ * Copyright (C) 2018 The "mysteriumnetwork/js-tequilapi" Authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,24 +15,18 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { validateMultiple } from '../validation'
+import { validateMultiple } from '../fmt/validation'
 
-export interface ConnectionStatisticsDTO {
-  duration: number,
-  bytesReceived: number,
-  bytesSent: number
+export interface ConnectCountDTO {
+  success: number,
+  fail: number,
+  timeout: number
 }
 
-export function parseConnectionStatisticsDTO (data: any): ConnectionStatisticsDTO {
-  validateMultiple('ConnectionStatisticsDTO', data, [
-    { name: 'duration', type: 'number' },
-    { name: 'bytesReceived', type: 'number' },
-    { name: 'bytesSent', type: 'number' }
+export function parseConnectionCountDTO (data: any): ConnectCountDTO {
+  validateMultiple('ConnectCountDTO', data, [
+    { name: 'success', type: 'number' },
+    { name: 'fail', type: 'number' }
   ])
-
-  return {
-    duration: data.duration,
-    bytesReceived: data.bytesReceived,
-    bytesSent: data.bytesSent
-  }
+  return { success: 1, fail: 2, timeout: 3 }
 }
