+++ src/consumer/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 The "mysteriumnetwork/js-tequilapi" Authors.
+ * Copyright (C) 2019 The "mysteriumnetwork/mysterium-vpn-js" Authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,13 +15,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { validate } from '../validation'
-
-export interface IdentityPayoutDTO {
-  ethAddress: string
-}
-
-export function parseIdentityPayoutDTO (data: any): IdentityPayoutDTO {
-  validate('IdentityPayoutDTO', data, { name: 'eth_address', type: 'string' })
-  return { ethAddress: data.eth_address }
-}
+export * from './consumer-location'
