+++ src/http/index.ts
@@ -15,27 +15,6 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Metrics } from './metrics'
-import { ProviderService } from './provider-service'
-import { ProviderSessions } from './provider-sessions'
-import { Publisher } from './publisher'
-import { QualityCalculator } from './quality-calculator'
-import { QualityLevel } from './quality-level'
-import { ServiceStatus } from './service-status'
-
-export * from './location'
-export * from './service-definition'
-export * from './service-info'
-export * from './service-list'
-export * from './service-request'
-export * from './service-session'
-
-export {
-  Metrics,
-  ProviderService,
-  ProviderSessions,
-  Publisher,
-  QualityCalculator,
-  QualityLevel,
-  ServiceStatus,
-}
+export * from './axios-adapter'
+export * from './interface'
+export * from './timeouts'
