+++ src/metric/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019 The "mysteriumnetwork/js-tequilapi" Authors.
+ * Copyright (C) 2019 The "mysteriumnetwork/mysterium-vpn-js" Authors.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,11 +15,4 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { validateArray } from '../validation'
-import { parseServiceInfoDTO, ServiceInfoDTO } from './service-info'
-
-export function parseServiceListDTO (responseData: any): ServiceInfoDTO[] {
-  validateArray('ServiceInfoDTO[]', responseData)
-
-  return responseData.map(parseServiceInfoDTO)
-}
+export * from './metrics-dto'
