+++ src/nat/nat-status.ts
