+++ src/tequilapi-error.ts
